+++ packages/react-sdk/CHANGELOG.md
@@ -4,7 +4,20 @@
 
 ### Minor Changes
 
-- Update xmtp-js to version 11 [`b121750`](https://github.com/xmtp/xmtp-web/commit/b121750f33f42ee40a9d3a5b4d196c433d1209e5)
+- Update xmtp-js to version 11 [`b121750`](https://github.com/xmtp/xmtp-web/commit/b121750f33f42ee40a9d3a5b4d196c433d1209e5)\
+
+## 1.3.6
+
+### Patch Changes
+
+- Rename `hasTopic` => `hasConversationTopic`
+- Add exports for `hasConversationTopic`, `saveConversation`, `setConversationUpdatedAt`, `updateConversation`, `updateConversationMetadata`, `deleteMessage`, `getLastMessage`, `saveMessage`, `updateMessage`, and `updateMessageMetadata`
+
+## 1.3.5
+
+### Patch Changes
+
+- Export some conversation helpers
 
 ## 1.3.4